+++ CHANGELOG.rst
@@ -17,6 +17,9 @@ of the list).
 
 3.0.2 (unreleased)
 ==================
+
+- Added test code to validate astrometric solutions from the database. [#294]
+
 - Fixed a bug in the 2D histogram code used for estimating shifts for
   catalog pre-matching. This may result in better matching. [#286]